# openapi.yaml
openapi: 3.0.3
info:
  title: AI-Based Documents Processing API
  version: 1.0.0
  description: Flask API for document upload, OCR, extraction, export, and review.
servers:
  - url: http://localhost:20373
paths:
...
  /api/documents:
    get:
      summary: List documents for the authenticated user
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 20
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: document_type
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of documents
  /api/documents/{documentId}:
    get:
      summary: Get a single document
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document payload
...